Added in line handlers, and got tests running.

+++ fixgrep-core/src/test/groovy/org/tools4j/fixgrep/FixGrepOnLargeFileTest.groovy
@@ -31,7 +31,7 @@ class FixGrepOnLargeFileTest extends Specification {
         assert assertTwoFilesAreEqual(actualOutputFile, expectedOutputFile)
 
         cleanup:
-        actualOutputFile.delete()
+        if(actualOutputFile.exists()) actualOutputFile.delete()
     }
 
     boolean assertTwoFilesAreEqual(final File actualFile, final BufferedReader expected) {
